Improved handling in scalaConfig

+++ R/rscala/R/scalaConfig.R
@@ -42,18 +42,19 @@ scalaConfig <- function(verbose=TRUE, reconfig=FALSE, download=character(0), req
     if ( !identical(reconfig,"live") && interactive() ) {
       while ( TRUE ) {
         cat(msg,"\n")
-        response <- toupper(trimws(readline(prompt="Would you like to install/update it now? [Y/n] ")))
+        response <- toupper(trimws(readline(prompt=paste0("Do you want to install here: ",installPath,"?  [Y/n] "))))
         if ( response == "N" ) return(FALSE)
         if ( response %in% c("Y","") ) return(TRUE)
       }
     } else FALSE
   }
+  dependsPath <- file.path(Sys.getenv("R_PACKAGE_DIR"),"systemRequirements")
   configPath <- file.path(installPath,"config.R")
-  if ( identical(reconfig,FALSE) && file.exists(configPath) && !download.java && !download.scala && !download.sbt ) {
-    if ( verbose ) cat(paste0("\nRead existing configuration file: ",configPath,"\n\n"))
+  if ( identical(reconfig,FALSE) && file.exists(configPath) && !download.java && !download.scala && !download.sbt && ( ! identical(Sys.getenv("R_INSTALL_PKG"),"rscala") ) ) {
+    if ( verbose ) cat(paste0("\nRead configuration file from rscala home: ",configPath,"\n\n"))
     source(configPath,chdir=TRUE,local=TRUE)
-    if ( is.null(config$format) || ( config$format < 4L ) || ( ! all(file.exists(c(config$javaHome,config$scalaHome,config$javaCmd,config$scalaCmd))) ) || ( is.null(config$sbtCmd) && require.sbt ) || ( ! is.null(config$sbtCmd) && ! file.exists(config$sbtCmd) ) ) {
-      if ( verbose ) cat("The 'config.R' is out-of-date.  Reconfiguring...\n")
+    if ( isConfigBroken(config, require.sbt) ) {
+      if ( verbose ) cat("The configuration is out-of-date.  Reconfiguring...\n")
       unlink(configPath)
       scalaConfig(verbose, reconfig, download, require.sbt)
     } else config
@@ -62,7 +63,7 @@ scalaConfig <- function(verbose=TRUE, reconfig=FALSE, download=character(0), req
     javaConf <- findExecutable("java","Java",installPath,javaSpecifics,verbose)
     if ( is.null(javaConf) ) {
       if ( verbose ) cat("\n")
-      consent2 <- offerInstall(paste0("Java and Scala are not found.")) 
+      consent2 <- offerInstall("Java and Scala are not found.")
       stopMsg <- "\n\n<<<<<<<<<<\n<<<<<<<<<<\n<<<<<<<<<<\n\nJava is not found!  Please run 'rscala::scalaConfig(download=\"java\")'\n\n>>>>>>>>>>\n>>>>>>>>>>\n>>>>>>>>>>\n"
       if ( consent2 ) {
         installSoftware(installPath,"java",verbose=verbose)
@@ -70,16 +71,16 @@ scalaConfig <- function(verbose=TRUE, reconfig=FALSE, download=character(0), req
         if ( is.null(javaConf) ) stop(stopMsg)
       } else {
         if ( identical(Sys.getenv("R_INSTALL_PKG"),"rscala") ) {
-          dependsPath <- file.path(Sys.getenv("R_PACKAGE_DIR"),"systemRequirements")
           installSoftware(dependsPath,"java",verbose=verbose)
           javaConf <- findExecutable("java","Java",dependsPath,javaSpecifics,verbose)
         }
         if ( is.null(javaConf) ) {
-          tmpdir <- tempdir()
+          tmpdir <- file.path(tempdir(),"rscala")
           installSoftware(tmpdir,"java",verbose=verbose)
           javaConf <- findExecutable("java","Java",tmpdir,javaSpecifics,verbose)
           if ( is.null(javaConf) ) stop(stopMsg)
-          else if ( verbose || interactive() ) cat(stopMsg)
+          stopMsg <- "<<<<<<<<<<\n<<<<<<<<<<\n<<<<<<<<<<\n\nJava was downloaded to a temporarly directory.  To avoid this, please run 'rscala::scalaConfig(download=\"java\")'\n\n>>>>>>>>>>\n>>>>>>>>>>\n>>>>>>>>>>\n"
+          cat(stopMsg)
         }
       }
       consent <- consent || consent2
@@ -89,7 +90,7 @@ scalaConfig <- function(verbose=TRUE, reconfig=FALSE, download=character(0), req
     scalaConf <- findExecutable("scala","Scala",installPath,scalaSpecifics2,verbose)
     if ( is.null(scalaConf) ) {
       if ( verbose ) cat("\n")
-      consent2 <- consent || offerInstall(paste0("Scala is not found.")) 
+      consent2 <- consent || offerInstall("Scala is not found.")
       stopMsg <- "\n\n<<<<<<<<<<\n<<<<<<<<<<\n<<<<<<<<<<\n\nScala is not found!  Please run 'rscala::scalaConfig(download=\"scala\")'\n\n>>>>>>>>>>\n>>>>>>>>>>\n>>>>>>>>>>\n"
       if ( consent2 ) {
         installSoftware(installPath,"scala",verbose=verbose)
@@ -97,16 +98,16 @@ scalaConfig <- function(verbose=TRUE, reconfig=FALSE, download=character(0), req
         if ( is.null(scalaConf) ) stop(stopMsg)
       } else {
         if ( identical(Sys.getenv("R_INSTALL_PKG"),"rscala") ) {
-          dependsPath <- file.path(Sys.getenv("R_PACKAGE_DIR"),"systemRequirements")
           installSoftware(dependsPath,"scala",verbose=verbose)
           scalaConf <- findExecutable("scala","Scala",dependsPath,scalaSpecifics2,verbose)
         }
         if ( is.null(scalaConf) ) {
-          tmpdir <- tempdir()
+          tmpdir <- file.path(tempdir(),"rscala")
           installSoftware(tmpdir,"scala",verbose=verbose)
           scalaConf <- findExecutable("scala","Scala",tmpdir,scalaSpecifics2,verbose)
           if ( is.null(scalaConf) ) stop(stopMsg)
-          else if ( verbose || interactive() ) cat(stopMsg)
+          stopMsg <- "<<<<<<<<<<\n<<<<<<<<<<\n<<<<<<<<<<\n\nScala was downloaded to a temporarly directory.  To avoid this, please run 'rscala::scalaConfig(download=\"scala\")'\n\n>>>>>>>>>>\n>>>>>>>>>>\n>>>>>>>>>>\n"
+          cat(stopMsg)
         }
       }
       consent <- consent || consent2
@@ -117,7 +118,7 @@ scalaConfig <- function(verbose=TRUE, reconfig=FALSE, download=character(0), req
     sbtConf <- findExecutable("sbt","SBT",installPath,sbtSpecifics,verbose)
     if ( is.null(sbtConf) && require.sbt ) {
       if ( verbose ) cat("\n")
-      consent2 <- consent || offerInstall(paste0("SBT is not found.")) 
+      consent2 <- consent || offerInstall("SBT is not found.")
       stopMsg <- "\n\n<<<<<<<<<<\n<<<<<<<<<<\n<<<<<<<<<<\n\nSBT is not found!  Please run 'rscala::scalaConfig(download=\"sbt\")'\n\n>>>>>>>>>>\n>>>>>>>>>>\n>>>>>>>>>>\n"
       if ( consent2 ) {
         installSoftware(installPath,"sbt",verbose=verbose)
@@ -128,16 +129,27 @@ scalaConfig <- function(verbose=TRUE, reconfig=FALSE, download=character(0), req
     }
     config <- c(config,sbtConf)
     if ( !consent && verbose ) cat("\n")
-    writeConfig <- consent || offerInstall(paste0("File '",configPath,"' is not found or is out-of-date."))
-    if ( writeConfig ) {
-      dir.create(installPath,showWarnings=FALSE,recursive=TRUE)
-      outFile <- file(configPath,open="w")
-      dump("config",file=outFile)
-      close(outFile)
-      if ( verbose ) cat(paste0("\nWrote configuration file: ",configPath,"\n\n"))
-    } else if ( verbose ) cat("\n")
+    writeConfig <- consent || offerInstall(paste0("Configuration file is not found (or is out-of-date)."))
+    if ( writeConfig ) writeConfigFile(configPath, config, verbose)
+    else if ( verbose ) cat("\n")
     config
   }
+}
+
+isConfigBroken <- function(config, require.sbt) {
+  result <- try(silent=TRUE,
+    ( is.null(config$format) || ( config$format < 4L ) || ( ! all(file.exists(c(config$javaHome,config$scalaHome,config$javaCmd,config$scalaCmd))) ) || ( is.null(config$sbtCmd) && require.sbt ) || ( ! is.null(config$sbtCmd) && ! file.exists(config$sbtCmd) ) )
+  )
+  inherits(result,"try-error") || result
+}
+
+writeConfigFile <- function(configPath, config, verbose) {
+  installPath <- dirname(configPath)
+  dir.create(installPath,showWarnings=FALSE,recursive=TRUE)
+  outFile <- file(configPath,open="w")
+  dump("config",file=outFile)
+  close(outFile)
+  if ( verbose ) cat(paste0("\nWrote configuration file: ",configPath,"\n\n"))
 }
 
 findExecutable <- function(mode,prettyMode,installPath,mapper,verbose=TRUE) {  ## Mimic how the 'scala' script finds Java.
@@ -203,6 +215,15 @@ findExecutable <- function(mode,prettyMode,installPath,mapper,verbose=TRUE) {  #
     conf <- tryCandidate(file.path(dependsPath,candidates))
     if ( ! is.null(conf) ) return(conf)    
   }
+  ###
+  label <- "rscala temporary directory"
+  regex <- sprintf("%s%s$",mode,if ( .Platform$OS.type == "windows" ) "(\\.exe|\\.bat)" else "")
+  installPath <- file.path(tempdir(),"rscala")
+  candidates <- list.files(installPath,paste0("^",regex),recursive=TRUE)
+  candidates <- candidates[grepl(sprintf("^%s/(.*/|)bin/%s",mode,regex),candidates)]
+  if ( length(candidates) > 1 ) candidates <- candidates[which.min(nchar(candidates))]
+  conf <- tryCandidate(file.path(installPath,candidates))
+  if ( ! is.null(conf) ) return(conf)
   NULL
 }
